Pandoc: Constrain width of some images
These images look bad if they take the full width of a line in a PDF
document.  Unfortunately, I have not found a possibility to constrain
the width of an image in Pandoc that would also work on GitHub.  Thus,
this commit patches the Markdown files before they are passed to Pandoc.